LayoutTests/fullscreen/full-screen-cancel-expected.txt

EVENT(webkitfullscreenchange)
EXPECTED (document.webkitIsFullScreen == 'true') OK
EXPECTED (document.webkitCurrentFullScreenElement == '[object HTMLElement]') OK
EVENT(webkitfullscreenchange)
EXPECTED (document.webkitIsFullScreen == 'false') OK
EXPECTED (document.webkitCurrentFullScreenElement == 'undefined') OK
END OF TEST
